+++ lipsum-dynamo.php
@@ -58,17 +58,20 @@ function lipnamo_install(){
 	
 	$charset_collate = $wpdb->get_charset_collate();
 	
-	$sql = "CREATE TABLE $table_name (
-		id bigint(20) NOT NULL AUTO_INCREMENT,
-		post_id bigint(20) NOT NULL,
-		time datetime DEFAULT '0000-00-00 00:00:00' NOT NULL,
-		PRIMARY KEY  (id)
-	) $charset_collate;";
-	
-	require_once(ABSPATH . 'wp-admin/includes/upgrade.php');
-	dbDelta($sql);
-	
-	add_option('lipnamo_db_version', LIPNAMO_VERSION);
+	$query = $wpdb->prepare('SHOW TABLES LIKE %s', $wpdb->esc_like($table_name));
+	if(!$wpdb->get_var($query) == $table_name){
+		$sql = "CREATE TABLE $table_name (
+					id bigint(20) NOT NULL AUTO_INCREMENT,
+					post_id bigint(20) NOT NULL,
+					time datetime DEFAULT '0000-00-00 00:00:00' NOT NULL,
+					PRIMARY KEY  (id)
+				) $charset_collate;";
+		
+		require_once(ABSPATH . 'wp-admin/includes/upgrade.php');
+		dbDelta($sql);
+		
+		add_option('lipnamo_db_version', LIPNAMO_VERSION);
+	}
 }
 
 add_action('plugins_loaded', 'lipnamo_update_db_check');