+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 1.1.0 - 2021-09-22
 
 * Changed: Add option to prevent deleting generated items but still cleanup database.
+* Fixed: Check database table exists before creating the new one.
 
 ## 1.0.0 - 2021-09-07